Updated view Logs button
Will now open the logs in a new tab, rather than the same tab. Added this because the only way to go back (if the log opened on the same page) would be to press the back button on the web browser. This button is not available on most smaller devices.